Refactor: DRY principle
