+++ tests/integration/helm/upgrade/n_minus_2/helm_upgrade_test.go
@@ -16,24 +16,15 @@
 package nminus2
 
 import (
-	"context"
-	"fmt"
-	"io/ioutil"
-	"os"
+	helmupgrade "istio.io/istio/tests/integration/helm/upgrade"
 	"path/filepath"
 	"testing"
 
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-
 	"istio.io/istio/pkg/test/env"
 	"istio.io/istio/pkg/test/framework"
-	"istio.io/istio/pkg/test/framework/components/cluster"
 	kubecluster "istio.io/istio/pkg/test/framework/components/cluster/kube"
 	"istio.io/istio/pkg/test/framework/image"
 	"istio.io/istio/pkg/test/helm"
-	kubetest "istio.io/istio/pkg/test/kube"
-	"istio.io/istio/pkg/test/scopes"
-	"istio.io/istio/pkg/test/util/retry"
 	helmtest "istio.io/istio/tests/integration/helm"
 	"istio.io/istio/tests/util/sanitycheck"
 )
@@ -52,8 +43,8 @@ global:
 `
 )
 
-// TestDefaultInPlaceUpgrades tests Istio installation using Helm with default options
-func TestDefaultInPlaceUpgrades(t *testing.T) {
+// TestDefaultInPlaceUpgradeFromTwoMinorReleases tests Istio installation using Helm with default options for Istio 1.(n-2)
+func TestDefaultInPlaceUpgradeFromTwoMinorReleases(t *testing.T) {
 	framework.
 		NewTest(t).
 		Features("installation.helm.default.upgrade").
@@ -64,11 +55,14 @@ func TestDefaultInPlaceUpgrades(t *testing.T) {
 			ctx.ConditionalCleanup(func() {
 				// only need to do call this once as helm doesn't need to remove
 				// all versions
-				deleteIstio(cs, h)
+				err := helmupgrade.DeleteIstio(cs, h)
+				if err != nil {
+					ctx.Fatalf("could not delete istio: %v", err)
+				}
 			})
 
-			overrideValuesFile := getValuesOverrides(ctx, defaultValues, gcrHub, nMinusTwoVersion)
-			installIstio(t, cs, h, overrideValuesFile)
+			overrideValuesFile := helmupgrade.GetValuesOverrides(ctx, defaultValues, gcrHub, nMinusTwoVersion)
+			helmupgrade.InstallIstio(t, cs, h, overrideValuesFile)
 			helmtest.VerifyInstallation(ctx, cs)
 
 			oldClient, oldServer := sanitycheck.SetupTrafficTest(t, ctx)
@@ -82,8 +76,8 @@ func TestDefaultInPlaceUpgrades(t *testing.T) {
 				ctx.Fatal(err)
 			}
 
-			overrideValuesFile = getValuesOverrides(ctx, defaultValues, s.Hub, s.Tag)
-			upgradeCharts(ctx, h, overrideValuesFile)
+			overrideValuesFile = helmupgrade.GetValuesOverrides(ctx, defaultValues, s.Hub, s.Tag)
+			helmupgrade.UpgradeCharts(ctx, h, overrideValuesFile)
 			helmtest.VerifyInstallation(ctx, cs)
 
 			newClient, newServer := sanitycheck.SetupTrafficTest(t, ctx)
@@ -92,95 +86,4 @@ func TestDefaultInPlaceUpgrades(t *testing.T) {
 			// now check that we are compatible with N-1 proxy with N proxy
 			sanitycheck.RunTrafficTestClientServer(t, oldClient, newServer)
 		})
-}
-
-// upgradeCharts upgrades Istio using Helm charts with the provided
-// override values file to the latest charts in $ISTIO_SRC/manifests
-func upgradeCharts(ctx framework.TestContext, h *helm.Helm, overrideValuesFile string) {
-	// Upgrade base chart
-	err := h.UpgradeChart(helmtest.BaseReleaseName, filepath.Join(helmtest.ChartPath, helmtest.BaseChart),
-		helmtest.IstioNamespace, overrideValuesFile, helmtest.HelmTimeout)
-	if err != nil {
-		ctx.Fatalf("failed to upgrade istio %s chart", helmtest.BaseChart)
-	}
-
-	// Upgrade discovery chart
-	err = h.UpgradeChart(helmtest.IstiodReleaseName, filepath.Join(helmtest.ChartPath, helmtest.ControlChartsDir, helmtest.DiscoveryChart),
-		helmtest.IstioNamespace, overrideValuesFile, helmtest.HelmTimeout)
-	if err != nil {
-		ctx.Fatalf("failed to upgrade istio %s chart", helmtest.DiscoveryChart)
-	}
-
-	// Upgrade ingress gateway chart
-	err = h.UpgradeChart(helmtest.IngressReleaseName, filepath.Join(helmtest.ChartPath, helmtest.GatewayChartsDir, helmtest.IngressGatewayChart),
-		helmtest.IstioNamespace, overrideValuesFile, helmtest.HelmTimeout)
-	if err != nil {
-		ctx.Fatalf("failed to upgrade istio %s chart", helmtest.IngressGatewayChart)
-	}
-
-	// Upgrade egress gateway chart
-	err = h.UpgradeChart(helmtest.EgressReleaseName, filepath.Join(helmtest.ChartPath, helmtest.GatewayChartsDir, helmtest.EgressGatewayChart),
-		helmtest.IstioNamespace, overrideValuesFile, helmtest.HelmTimeout)
-	if err != nil {
-		ctx.Fatalf("failed to upgrade istio %s chart", helmtest.EgressGatewayChart)
-	}
-}
-
-// installIstio install Istio using Helm charts with the provided
-// override values file and fails the tests on any failures.
-func installIstio(t *testing.T, cs cluster.Cluster,
-	h *helm.Helm, overrideValuesFile string) {
-	helmtest.CreateIstioSystemNamespace(t, cs)
-
-	// Install base chart
-	err := h.InstallChart(helmtest.BaseReleaseName, helmtest.BaseChart+helmtest.TarGzSuffix,
-		helmtest.IstioNamespace, overrideValuesFile, helmtest.HelmTimeout)
-	if err != nil {
-		t.Errorf("failed to install istio %s chart", helmtest.BaseChart)
-	}
-
-	// Install discovery chart
-	err = h.InstallChart(helmtest.IstiodReleaseName, filepath.Join(helmtest.ControlChartsDir, helmtest.DiscoveryChart)+helmtest.TarGzSuffix,
-		helmtest.IstioNamespace, overrideValuesFile, helmtest.HelmTimeout)
-	if err != nil {
-		t.Errorf("failed to install istio %s chart", helmtest.DiscoveryChart)
-	}
-
-	helmtest.InstallGatewaysCharts(t, cs, h, helmtest.TarGzSuffix, overrideValuesFile)
-}
-
-// deleteIstio deletes installed Istio Helm charts and resources
-func deleteIstio(cs cluster.Cluster, h *helm.Helm) error {
-	scopes.Framework.Infof("cleaning up resources")
-	if err := h.DeleteChart(helmtest.EgressReleaseName, helmtest.IstioNamespace); err != nil {
-		return fmt.Errorf("failed to delete %s release", helmtest.EgressReleaseName)
-	}
-	if err := h.DeleteChart(helmtest.IngressReleaseName, helmtest.IstioNamespace); err != nil {
-		return fmt.Errorf("failed to delete %s release", helmtest.IngressReleaseName)
-	}
-	if err := h.DeleteChart(helmtest.IstiodReleaseName, helmtest.IstioNamespace); err != nil {
-		return fmt.Errorf("failed to delete %s release", helmtest.IngressReleaseName)
-	}
-	if err := h.DeleteChart(helmtest.BaseReleaseName, helmtest.IstioNamespace); err != nil {
-		return fmt.Errorf("failed to delete %s release", helmtest.BaseReleaseName)
-	}
-	if err := cs.CoreV1().Namespaces().Delete(context.TODO(), helmtest.IstioNamespace, metav1.DeleteOptions{}); err != nil {
-		return fmt.Errorf("failed to delete istio namespace: %v", err)
-	}
-	if err := kubetest.WaitForNamespaceDeletion(cs, helmtest.IstioNamespace, retry.Timeout(helmtest.RetryTimeOut)); err != nil {
-		return fmt.Errorf("wating for istio namespace to be deleted: %v", err)
-	}
-
-	return nil
-}
-
-func getValuesOverrides(ctx framework.TestContext, valuesStr, hub, tag string) string {
-	workDir := ctx.CreateTmpDirectoryOrFail("helm")
-	overrideValues := fmt.Sprintf(valuesStr, hub, tag)
-	overrideValuesFile := filepath.Join(workDir, "values.yaml")
-	if err := ioutil.WriteFile(overrideValuesFile, []byte(overrideValues), os.ModePerm); err != nil {
-		ctx.Fatalf("failed to write iop cr file: %v", err)
-	}
-
-	return overrideValuesFile
 }
